BD/requirements.txt
pymongo==4.6.1
neo4j==5.16.0
pydantic>=2.6.0
python-dotenv==1.0.0
fastapi==0.104.0
uvicorn==0.24.0